big change

+++ Pages/Admin/delete_project.php
@@ -1,6 +1,14 @@
 <?php
-require '../../Controllers/accessDatabase.php';
+$servername = "127.0.0.1";
+$username = "root";
+$password = "";
+$dbname = "fcmDB";
 
+$conn = new mysqli($servername, $username, $password, $dbname);
+
+if ($conn->connect_error) {
+    die("Connection failed: " . $conn->connect_error);
+}
 $projectid = $_POST['projectid'];
 $sql = "DELETE FROM project WHERE projectid='$projectid'";
